refactor: split shelf_detail.rs into a shelf_detail/ module (#1253)
shelf_detail.rs had grown to 857 lines, over the ~800-line file cap.
Split it into shelf_detail/{mod,body,header,add_books_modal}.rs by
sub-topic, mirroring the db/src/books/ and book_detail/ directory
patterns. Pure move: no rsx or logic changes, same tests (relocated
alongside the code they cover where applicable).

book_detail/mod.rs was re-checked and is 564 lines, already under the
cap from prior unrelated work, so it was left untouched.
